worker/requirements.txt
kafka-python
nltk
pymysql
cryptography